update elFinder to be jQuery 4 friendly

+++ external/elFinder/php/elFinder.class.php
@@ -2015,7 +2015,7 @@ class elFinder
                     return $a404;
                 }
                 $GLOBALS['elFinderTempFiles'][$src] = true;
-                
+
                 unset($file['file']);
                 $file['read'] = true;
                 $file['size'] = filesize($src);
@@ -4394,7 +4394,7 @@ var go = function() {
     protected function utime()
     {
         $time = explode(" ", microtime());
-        return (double)$time[1] + (double)$time[0];
+        return (float)$time[1] + (float)$time[0];
     }
 
     /**
